Fix gpgpu module circular reference
