+++ modules/gpgpu/src/operations/cpu/arithmetic.ts
@@ -5,10 +5,10 @@
 import {OperationHandler} from '../../operation/operation';
 import {Expression} from '../../utils/expression';
 import {
-  ArithmeticOperationInputs,
-  ArithmeticOp,
-  ARITHMETIC_OPERATIONS
-} from '../arithmetic-operation';
+  ARITHMETIC_OPERATIONS,
+  type ArithmeticOperationInputs,
+  type ArithmeticOp
+} from '../arithmetic-expression';
 import {getValueAtRow} from './common';
 
 export const arithmetic: OperationHandler<ArithmeticOperationInputs> = ({
